64 add auto grader script

Npm maintenance is messing with CI

+++ Phase-2/islands/Admin.tsx
@@ -57,10 +57,15 @@ export default function Admin() {
 				.find((row) => row.startsWith("authToken="))
 				?.split("=")[1];
 
-			const response = await fetch(`${APIBaseURL}api/users`, {
+			if (!authToken) {
+				setIsAuthorized(false);
+				return;
+			}
+
+			const response = await fetch(`${APIBaseURL}users`, {
 				headers: {
 					"Content-Type": "application/json",
-					"X-authorization": authToken || "",
+					"X-Authorization": authToken,
 				},
 			});
 			if (!response.ok) throw new Error("Failed to fetch users.");
@@ -78,14 +83,20 @@ export default function Admin() {
 				.find((row) => row.startsWith("authToken="))
 				?.split("=")[1];
 
-			const response = await fetch(`${APIBaseURL}api/users`, {
+			if (!authToken) {
+				setIsAuthorized(false);
+				return;
+			}
+
+			const response = await fetch(`${APIBaseURL}users`, {
 				method: "POST",
 				headers: {
 					"Content-Type": "application/json",
-					"X-authorization": authToken || "",
+					"X-Authorization": authToken,
 				},
 				body: JSON.stringify(newUser),
 			});
+
 			if (!response.ok) throw new Error("Failed to create user.");
 			fetchUsers(); // Refresh user list
 			setShowCreateForm(false); // Close the create form
@@ -110,11 +121,16 @@ export default function Admin() {
 				.find((row) => row.startsWith("authToken="))
 				?.split("=")[1];
 
-			const response = await fetch(`${APIBaseURL}api/users/${user.username}`, {
+			if (!authToken) {
+				setIsAuthorized(false);
+				return;
+			}
+
+			const response = await fetch(`${APIBaseURL}users/${user.username}`, {
 				method: "PUT",
 				headers: {
 					"Content-Type": "application/json",
-					"X-authorization": authToken || "",
+					"X-Authorization": authToken,
 				},
 				body: JSON.stringify(user),
 			});
@@ -144,11 +160,16 @@ export default function Admin() {
 				.find((row) => row.startsWith("authToken="))
 				?.split("=")[1];
 
-			const response = await fetch(`${APIBaseURL}api/users/${userToDelete}`, {
+			if (!authToken) {
+				setIsAuthorized(false);
+				return;
+			}
+
+			const response = await fetch(`${APIBaseURL}users/${userToDelete}`, {
 				method: "DELETE",
 				headers: {
 					"Content-Type": "application/json",
-					"X-authorization": authToken || "",
+					"X-Authorization": authToken,
 				},
 			});
